+++ variables-naming.tf
@@ -23,3 +23,9 @@ variable "custom_network_security_group_name" {
   type        = string
   default     = null
 }
+
+variable "custom_network_watcher_flow_log_name" {
+  description = "Network watcher flow log name."
+  type        = string
+  default     = null
+}